+++ test/test_parse_rules.rb
@@ -10,9 +10,9 @@ describe 'Crass::Parser' do
   make_my_diffs_pretty!
   parallelize_me!
 
-  describe '#parse_stylesheet' do
+  describe '#parse_rules' do
     def parse(*args)
-      CP.parse_stylesheet(*args)
+      CP.parse_rules(*args)
     end
 
     behaves_like 'parsing a list of rules'
@@ -20,23 +20,21 @@ describe 'Crass::Parser' do
     it 'with CDO/CDC before rule' do
       tree = parse(" <!-- --> {")
 
-      assert_equal 4, tree.size
+      assert_equal 2, tree.size
+      assert_tokens " ", tree[0]
 
-      tree[0..2].each do |node|
-        assert_equal :whitespace, node[:node]
-      end
-
-      rule = tree[3]
+      rule = tree[1]
       assert_equal :style_rule, rule[:node]
       assert_equal [], rule[:children]
 
       selector = rule[:selector]
       assert_equal :selector, selector[:node]
-      assert_equal "", selector[:value]
-      assert_equal [], selector[:tokens]
+      assert_equal "<!-- -->", selector[:value]
+      assert_tokens "<!-- --> ", selector[:tokens], 1
     end
 
     it 'followed by CDC' do
+      # TODO: This should be a parse error.
       tree = parse("div {} -->")
 
       assert_equal 2, tree.size